FrechetDist.jl / Low level details

Miscellaneous

We cover here a few low level internal functions. This is probably

only of interest if you are trying to understand the source coe of the package.

Functions documentation

FrechetDist.extract_param_inner — Function

extract_param_inner

Takes the sequence of events (i.e., sequence of points along the curve), and outputs a corresponding sequence of real numbers, where the ith nubmer if the distance of the ith point in sequence from the beginning of the curve. Here, distance means the total distance of the subcurve from the start point, to the current point.

FrechetDist.extract_refined_polygon — Function

extract_refined_polygon( P, pes, num_points_to_add, delta )

Refine edges along which pes (the sequence of matching points
along P is not monotone, so that the returned polygon has
vertices, which hopefully force the Frechet morphing computed in
the next iteration to be monotone.

Arguments

P : the polygo pes: Sequence of events long the polygon P num_points_to_add: how many points to add between points of non-monotonicity. The default is 1. Having a larger value speeds up the process of getting rid of non-monotonicity, but the polygons grow faster.

FrechetDist.frechet_c_mono_approx_subcurve — Function

frechet_c_mono_approx_subcurve

Approximates the Frechet distance between a curve (P) and subcurbe (Q). Here, Q vertices are the vertices of P specifieid by pindices. That is Q[ i ] =P[ pindices[ i ]].

FrechetDist.frechet_mono_via_refinement — Function

frechet_mono_via_refinement( P, Q, approx )

Computes the "true" monotone Frechet distance between P and Q,
using the ve_r algorithm. It does refinement, to add vertices if
needed to get monotonicity. Note, that there is an eps parameter -
for real inputs you can set it quite small. In any case, in the
worst case it only approximates the Frechet (monotone)
distance. It returns a morphing, and a boolean that is true if the
result is the true Frechet distance.

Observe, that this function might be too slow if the two curves
are huge, since it does not simplify them before computing the
distance.

Returns

Returns a 4-tuple ( m, f, P, Q ): m: Morphing realizing result. f: Is distance returned is the exact continuous Frechet distance. P: Refined first input polygon Q: Refined second input polygon

'out' if specified returns the middle morphings used in computing the solution.

FrechetDist.frechet_ve_r_mono_compute — Function

frechet_ve_r_mono_compute

Computing the ver frechet distance, and then monotonize it. No guarentee as far as the quality of the distance this realizes...

FrechetDist.Morphing_from_prm — Function

Morphing_from_prm( prm, P, Q )

Construct a morphing from a parameterization of the two polygons P
and Q.

FrechetDist.add_points_along_seg — Function

add_points_along_seg

Arguments

seg: t[1], t[2], .... , t[end]

 t[...]: Time stamps for points on the segment (all between 0 and 1)

Details

Add points along seg at time stamps specified by t, and add points in the middle between them (i.e., refining segment). These new points are added the output polygon pout.

Note, that one can set nmidpoints to be larger than 1, but it does not seem to be necessary/benefitial from the experimentation we did.

Output: Sorted( t[1]/2, t[1], (t[1]+t[2])/2, t[2],...)

FrechetDist.EID — Function

EID

A function that encodes a matching "edge" betwee two polygons.

FrechetDist.integral — Function

intgral

returns the integral of the function f, for the value x. This integral was computed by sage math.

FrechetDist.SweepDist_segs — Function

SweepDist_segs

Finally, compute the CDTW between two (oriented) segmets $p_a p_b$ and $q_a q_b$. Its simply the sum of teh two oriented integrals of the distance.

FrechetDist.DistFunc_verify_integral — Function

DistFunc_verify_integral

Verifies the intergral function presented above. Since we got the integral of the function f from software (i.e., sagemath), and I was too lazy to verify it by hand, I did a numerical verification.

FrechetDist.frechet_offsets — Function

frechet_offests

Given P, and subset of indices of P defining a subpolygon, computes a -approximation to the Frechet distance between every edge of the subpolygon they define, and corresponding portion of P. A fast, and probably decent approximation to the optimal morhping. Finally, set for every vertex of the subcurve, the frechet distance of this subcurve. Thus, providing a fast estimates of the offsets.

FrechetDist.SweepDist_segs_p — Function

SweepDist_segs_p

Return the oriented Sweep distance (i.e., continuous dynamic time wrapping) distance between two oriented segments. To this end, we compute the distance function, and then compute its integrval.

FrechetDist.integral_interval — Function

inegral_interval

Return the integral of f over an interval.

FrechetDist.SweepDist_lb_compute — Function

SweepDist_lb_compute

Compute a lower bound on the Sweep distance between the two curves P and Q. Current implementation is silly, and very similar to the discrete Frechet calculation.