/ Computational Geometry Primitives

The package provides some basic geometric types in the cg submodule.

Package Features

Provides types: Point, Segment, Line, Polygon.

Basic geometric types

Functions

FrechetDist.cg.BBox_init — Method

BBox_init

FrechetDist.cg.BBox_max_dist_in_dim — Method

BBox_extent_in_dim

The maximum distance between two points in this dimension, that

belongs to the two bounding boxes.

FrechetDist.cg.segs_match_price — Method

segs_match_price

The price of matching the edge pa-pb to the edge qa-qb.